衍佛光

+++ Core/Battle/BattleFighter.h
@@ -1128,12 +1128,12 @@ private:
 
     float auralRate;
     UInt8 auralLast;
-    UInt8 auraValue;
+    float auraValue;
     inline void setBuddhaLight(float v, UInt8 l) { auralRate = v; auralLast = l; }
     inline UInt8 getBuddhaLightLast() { return auralLast; }
     inline float getBuddhaLightRate() { return auralRate; }
-    inline void setBuddhaLightAura(UInt8 value) { auraValue = value; }
-    inline UInt8 getBuddhaLightAura() { return auraValue; }
+    inline void setBuddhaLightAura(float value) { auraValue = value; }
+    inline float getBuddhaLightAura() { return auraValue; }
 
     std::vector<GData::SkillItem> _passiveSkillDeadFake100;
     std::vector<GData::SkillItem> _passiveSkillDeadFake;
